// lib/woodSpecies.generated.ts
export interface SpeciesPhoto {
  fileId: string
  name: string
  width: number | null
  height: number | null
}